pyproject.toml
[project]
name = "mood-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
